Move old sound code to the dmg module
